network: vasprintf returns an int.

+++ src/network/io.c
@@ -541,7 +541,7 @@ ssize_t __net_vaPrintf( vlc_object_t *p_this, int fd, const v_socket_t *p_vs,
     char    *psz;
     int      i_ret;
 
-    size_t i_size = vasprintf( &psz, psz_fmt, args );
+    int i_size = vasprintf( &psz, psz_fmt, args );
     if( i_size == -1 )
         return -1;
     i_ret = __net_Write( p_this, fd, p_vs, (uint8_t *)psz, i_size ) < i_size
